+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/NoopStatementDecorator.java
@@ -15,13 +15,12 @@
 package com.ericsson.bss.cassandra.ecchronos.application;
 
 import com.datastax.driver.core.Statement;
+import com.ericsson.bss.cassandra.ecchronos.application.config.Config;
 import com.ericsson.bss.cassandra.ecchronos.connection.StatementDecorator;
-
-import java.util.Properties;
 
 public class NoopStatementDecorator implements StatementDecorator
 {
-    public NoopStatementDecorator(Properties properties) // NOPMD
+    public NoopStatementDecorator(Config config) // NOPMD
     {
         // Nothing to do
     }